perf(engine): let the reductions table say it has no negative entry
reductions[i] is int(2872 / 128.0 * log(i)) -- a scaled logarithm, 0 through
124 across the whole table, and never negative. As int that is a fact the
compiler cannot use, so the `/ 512` in reduction() has to allow for a negative
dividend and pays the round-toward-zero correction on every move the search
reduces. u16 carries the range instead, both factors of the product are then
provably non-negative, and the shift stands alone:

  clang 22.1.8, x86-64-avx2 -O3, search.cpp:2078

    before                          after
    imul $0xc5,%edx,%ecx            imul $0xc5,%edx,%ecx
    lea  0x1ff(%rcx),%esi           shr  $0x9,%ecx
    test %ecx,%ecx
    cmovns %ecx,%esi
    sar  $0x9,%esi

The table also halves, 1024 bytes to 512; nothing here measures that.

  valgrind --tool=callgrind, x86-64-avx2, clang -O3, marginal between depth 12
  and depth 16 on one middlegame position, 71,682 nodes, node counts equal

    engine-side Ir per node   -3.58, all of it inside Worker::search

  tests/perfbudget.sh --pgo --comp clang --arch x86-64-avx2 HEAD worktree,
  node counts equal at 182,697, startup within 51 Ir

    search Ir   -564,381   -0.0477%   PASS

The gcc cells were not run for this change, so the grid is one compiler wide
where the commit before it is four cells. No cycle or wall-clock number is
quoted: four other agents were building on this box throughout.

  tests/signature.sh 2516158, x86-64-avx512icl clang   exit 0
  tests/reprosearch.sh, 332,525 nodes                  exit 0
  tests/perft.sh, standard and Chess960                exit 0
  python3 tests/instrumented.py --none, 75 tests       exit 0

No functional change

+++ src/engine/search.h
@@ -568,7 +568,13 @@ class Worker {
     usize threadIdx, numaThreadIdx, numaTotal;
 
     // Reductions lookup table initialized at startup
-    std::array<int, MAX_MOVES> reductions;  // [depth or moveNumber]
+    //
+    // Unsigned because the table holds a scaled logarithm and has no negative
+    // entry, and because that is what the type has to say for the compiler to
+    // know it. reduction() divides their product by 512; as int it cannot rule
+    // out a negative dividend, so the shift comes with the three-instruction
+    // round-toward-zero correction on every move the search reduces.
+    std::array<u16, MAX_MOVES> reductions;  // [depth or moveNumber]
 
     // The main thread has a SearchManager, the others have a NullSearchManager.
     //
